+++ jdbc-yugabytedb/src/main/java/com/yugabyte/ysql/ClusterAwareLoadBalancer.java
@@ -1,4 +1,4 @@
-package org.postgresql.jdbc.yugabyte;
+package com.yugabyte.ysql;
 
 import java.sql.*;
 import java.util.*;
